book store app

+++ app/src/main/java/com/example/bookstoreapp/data/InventoryItem.kt
@@ -5,7 +5,8 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "inventory_items")
 data class InventoryItem(
-    @PrimaryKey(autoGenerate = true) val id: Long = 0,
+    @PrimaryKey(autoGenerate = true)
+    val id: Int = 0,
     val name: String,
     val category: String,
     val quantity: Int,
